Update imports and classnames after split
